Move rsa tests to their own file
